refactor german-agent project structure
