+++ backend/german-agent-microservice/legacy/res/data/a2_skript_gr0.md
