+++ backend/german-agent-microservice/legacy/src/document_handler.py
