+++ backend/german-agent-microservice/legacy/src/tool_master.py
