+++ backend/german-agent-microservice/src/agent_service/agent/agent.py
@@ -3,17 +3,18 @@ from pathlib import Path
 from llm import LLMBedrock
 from typing import List
 from agent_step import AgentStep
-from agent_service.utils.prompt.react_prompt import REACT_PROMPT
-from agent_service.utils.prompt.prompt_builder import PromptBuidler
+from agent_service.prompts.react_prompt import REACT_PROMPT
+from agent_service.prompts.prompt_builder import PromptBuidler
 from agent_service.tools.tool_executor import ToolExecutor
-from agent_service.utils.parsers.agent_answer_parser import AnswerParser
-from agent_service.utils.parsers.agent_step_parser import StepParser
+from agent_service.parsers.agent_answer_parser import AnswerParser
+from agent_service.parsers.agent_step_parser import StepParser
+
 class Agent:
-    CONFIG_PATH = Path("/Users/ali/Desktop/code/aika/backend/german_agent_microservice/src/agent_service/core/llm-config.ini")
+    CONFIG_PATH = Path("/backend/german_agent_microservice/src/core/llm-config.ini")
     def __init__(self) -> None:
         self.parse_config()
         self.reasoning_trace : List[AgentStep] = []
-        self.prompt = PromptBuidler(REACT_PROMPT)
+        self.prompt = PromptBuidler(REACT_PROMPT) # variables: tools, tool_names, input, reasoning_trace
         self.tool_executor = ToolExecutor()
         self.answer_parser = AnswerParser()
         self.step_parser = StepParser()
