+++ backend/german-agent-microservice/src/agent_service/agent/agent_step.py
@@ -1,5 +1,6 @@
 from pydantic import BaseModel, field_validator, Field
 from typing import List, Any
+
 class AgentStep(BaseModel):
     thought:str
     tool_names:List[str]
