+++ backend/german-agent-microservice/src/agent_service/parsers/agent_answer_parser.py
