+++ backend/german-agent-microservice/src/agent_service/parsers/agent_step_parser.py
