+++ backend/german-agent-microservice/src/agent_service/prompts/prompt_builder.py
