+++ backend/german-agent-microservice/src/agent_service/prompts/react_prompt.py
