+++ backend/german-agent-microservice/src/agent_service/prompts/reading_generation_prompt.py
