+++ backend/german-agent-microservice/src/agent_service/tools/tool.py
