+++ backend/german-agent-microservice/src/agent_service/tools/web_search_tool.py
