+++ backend/german-agent-microservice/src/main.py
@@ -1,3 +1,4 @@
+#### FOR DEBUGGING PURPOSES
 from agent_service.agent.llm import LLMBedrock
 from agent_service.tools.reading_generation_tool import ReadingGenerator
 import logging
