--- backend/german_agent_microservice/tests/agent_service/agent/config.py
@@ -1,1 +1,0 @@
-SRC="../../../src/"
